requirements.txt
fastapi
uvicorn[standard]
numpy
Pillow
opencv-python
tensorflow
openai
torch
torchvision
gradio
redis
requests
rapidfuzz
